feat: optionally pass type to column (#79)
Closes #78.

### Summary of Changes

For the sake of consistency it is now possible to pass the type of a
`Column` in the constructor. This also improves performance, for example
when we call `get_column` on a `Table`. In that case we already know the
type of the column anyway, so there's no reason to infer it again.

---------

Co-authored-by: lars-reimann <lars-reimann@users.noreply.github.com>