feat: return new model when calling `fit` (#91)
Closes #69.

### Summary of Changes

The `fit` method of classifiers/regressors now returns a new (fitted)
classifier regressor. The receiver of the method call is not changed
anymore. This is consistent with the methods on the `Table` class and
other data containers. Furthermore, `fit` is now a pure function, which
works better in notebooks and our own [execution
strategy](https://arxiv.org/abs/2302.14556).

---------

Co-authored-by: lars-reimann <lars-reimann@users.noreply.github.com>