feat: specify features and target when creating a `TaggedTable` (#114)
Closes #27 .

### Summary of Changes

A user can now optionally specify the `features` of a `TaggedTable`
explicitly. If the features are not specified, all columns except the
target are considered features.

The `predict` method of classifiers/regressors only uses the features
for prediction. The other columns are still included in the output,
however. This is, for example, useful to include an ID column in the
table created by `predict`.

---------

Co-authored-by: lars-reimann <lars-reimann@users.noreply.github.com>