feat: move existing ML solutions into `safeds.ml.classical` package (#213)
Closes #210.

### Summary of Changes

We eventually want to support neural networks as well, which are
supposed to reside in the `safeds.ml.neural` package. We do not want to
have packages with different semantics (classification and regression as
tasks, neural networks as a subset of ML methods) on the same package
level, however. Because of this, we moved the existing methods into the
`safeds.ml.classical` package.

---------

Co-authored-by: megalinter-bot <129584137+megalinter-bot@users.noreply.github.com>