fix: selectively ignore one warning instead of all warnings (#235)
### Summary of Changes

When we tried to test whether a specific warning was issued, we
discovered that these tests failed if all tests were run but passed if
only the test for the warning was run. This happened because one module
basically disabled all warnings when it was loaded. This behavior is now
removed and instead one specific warning is disabled for the one call
that may create it.