feat: replace prefix `n` with `number_of` (#250)
### Summary of Changes

* Rename `n_columns` to `number_of_columns`
* Rename `n_rows` to `number_of_rows`

Reason: The prefix `n` is commonly used instead of `number_of`. However,
we also need to add a parameter to `AdaBoost` to set the maximum number
of learners (#171). Calling that `maximum_n_learners` doesn't look
nearly as readable to me as `maximum_number_of_learners`. And if we
don't use `n` to replace `number_of` here, we shouldn't use this
anywhere.

---------

Co-authored-by: megalinter-bot <129584137+megalinter-bot@users.noreply.github.com>