feat: set kernel of support vector machine (#350)
Closes #172.

A new parameter kernel: SupportVectorMachineKernel was added to the
initializer of SupportVectorMachine classes of both the classifier and
the regressor alongside getters for the parameters. It was later passed
as the kernel of the wrapped scikit-learn model in the fit method.
The SupportVectorMachineKernel was created as an abstract base class
from which the signatures of the methods in the nested subclasses of
Kernel were the same.
Tests were adjusted to test the functionality of the kernel parameter.