feat: add suffixes to models to indicate their task (#588)
### Summary of Changes

Add the suffix `Classifier` to all models for classification and
`Regressor` to all models for regression. While being longer, this
naming has many advantages:

* Better **readability**: Several models have variants for
classification and regressions. Previously, both had the same name, so
to understand which one was used in the code, imports had to be checked.
* Better **auto-completion**: Now users can simply write `Classifier` or
`Regressor` to get a list of all suitable models.
* Better **understandability**: Now it's obvious, that logistic
regression is used for classification.

---------

Co-authored-by: megalinter-bot <129584137+megalinter-bot@users.noreply.github.com>