feat: remove output conversions (#792)
Closes partially #732

### Summary of Changes

Output conversions are the exact inversion of the input conversion, so
there is no need to specify them again. Now, a neural network only takes
an input conversion and a list of layers. This also gets rid of several
errors that could occur if input and output conversions did not fit
together.

In a later PR, the input conversion will also be removed, since they
mirror datasets.

---------

Co-authored-by: megalinter-bot <129584137+megalinter-bot@users.noreply.github.com>